Bunch of changes to the keymap, but this'll be squashed
* moved DASHES to separate key from hyphen, so I can type -- in
  commandline flags
* set modifier guard in DOT_DOTS so I can type >>> safely
* moved tap dance and muse mode code into separate files
* minor reformats
