+++ keyboards/planck/keymaps/3geek14/.gitignore
@@ -1,1 +1,1 @@
-/user-includes.h
+/user_includes.h
