+++ keyboards/planck/keymaps/3geek14/tap_dance.h
@@ -15,27 +15,12 @@
  */
 
 #pragma once
+#include "quantum.h"
 
-#ifdef AUDIO_ENABLE
-#    define STARTUP_SONG SONG(PLANCK_SOUND)
-#endif
+#define DOT_DOTS TD(TD_DOT_DOTS)
+#define DASHES TD(TD_DASHES)
 
-#define MIDI_BASIC
-#define HOLD_ON_OTHER_KEY_PRESS_PER_KEY
-#define DOUBLE_TAP_SHIFT_TURNS_ON_CAPS_WORD
-#define TAP_CODE_DELAY 10
-#define TAPPING_TERM 175
-
-// Unicode support for WinCompose with Right Alt as compose key
-#define UNICODE_SELECTED_MODES UNICODE_MODE_WINCOMPOSE
-#define UNICODE_KEY_WINC KC_RALT
-
-// Reduces firmware size and limits to 8 layers
-#define LAYER_STATE_8BIT
-
-#if __has_include("user_includes.h")
-#    include "user_includes.h"
-#else
-#    define EMAIL_STRING ""
-#    define SIGNATURE_STRING ""
-#endif
+enum tap_dance_declarations {
+    TD_DOT_DOTS,
+    TD_DASHES,
+};
